fix: reverting to original implementation style
