+++ libs/genai/tests/unit_tests/test_llms.py
@@ -22,6 +22,7 @@ def test_tracing_params() -> None:
         "ls_provider": "google_genai",
         "ls_model_type": "llm",
         "ls_model_name": MODEL_NAME,
+        "ls_temperature": 0.7,
     }
     assert llm.metadata is not None
     assert llm.metadata["lc_versions"]["langchain-google-genai"] == __version__
